test: triage the 10 tests/runtime/ files flagged by #158
Of the 10 candidate files #158 listed as importing persona controller/
state-machine/persona modules, 5 were already migrated to
tests/personas/<persona>/ between the issue's filing and now
(random-movement-ticks, complex-motivation-z3, actor-motivation-combat,
spatial-field-aware-actor-decision, director-artifact-seeding -- all
now live under tests/personas/actor/ or tests/personas/director/).

Of the remaining 5:

- design-spend-ledger.test.js -> migrated to
  tests/personas/allocator/allocator-design-spend-ledger.test.js.
  Every assertion is about buildDesignSpendLedger, an Allocator-owned
  function; Configurator/Director are injected collaborators exactly
  as production wires them, not the subject under test.

- affinity-field-per-tick.test.js -> split. One test drives a real
  runtime through real ticks and reads core-ts state directly
  (orchestration); the other two call createModeratorPersona().advance()
  directly with no runtime involved (pure Moderator behavior). They
  shared no helper functions, so the split was mechanical: the
  Moderator-only tests moved to
  tests/personas/moderator/moderator-plan-tick-close.test.js, the
  runtime-integration test and its helpers stay put.

- runtime-moderator-affinity-actions.test.js, runtime-budget-receipt.test.js,
  actor-proposal-replay.test.js -> left in tests/runtime/, each with a
  one-line header recording the triage decision. All three assert on
  what the RUNTIME does when wired to a real persona (applies actions
  to core, delegates receipt pricing, maps proposals to core actions),
  not on a persona's own decision output in isolation -- the
  orchestration/e2e half of the guard's stated exception, now written
  down instead of implicit.

All 7 touched/created files plus the layout guard
(tests/architecture/persona-test-layout.test.js) pass.

Fixes #158
