+++ tests/runtime/affinity-field-per-tick.test.js
@@ -1,6 +1,5 @@
 /**
- * AM.7 / AM.3b — the affinity field follows its source, and the MODERATOR closes
- * the tick (closes F7).
+ * AM.7 — the affinity field follows its source (closes F7).
  *
  * `computeAffinityField()` was called exactly once, during setup
  * (runner/core-setup.mjs). Actors then moved for the rest of the run while their
@@ -8,10 +7,12 @@
  * is decoration, and every consumer of it (tile visuals, the aura bridge, the
  * gameplay renderer) was reading a snapshot of tick 0.
  *
- * The recompute is deliberately NOT a second unconditional glue call. Advancing
- * the tick and recomputing the field describe the same instant, so both are one
- * Moderator decision (`plan_tick_close`), which also closes the AM.3b weakness:
- * before it, glue advanced the tick and the Moderator never said to.
+ * #158 — this is the runtime/core-ts integration counterpart to AM.3b (the
+ * MODERATOR's `plan_tick_close` decision to recompute the field, which used to
+ * live in this same file): that half calls the persona directly with no runtime
+ * involved and moved to tests/personas/moderator/moderator-plan-tick-close.test.js.
+ * This test stays here because it drives a real runtime through real ticks and
+ * reads core-ts state directly — orchestration, not persona-alone behavior.
  */
 "use strict";
 
@@ -137,61 +138,6 @@ test("the affinity field changes across ticks as its source actor moves", async 
     early,
     "the field must follow the actor. Identical fingerprints across a run in which the actor "
       + `moved ${earlyPos} -> ${latePos} mean it is still the setup-time snapshot (F7).`,
-  );
-});
-
-// ---------------------------------------------------------------------------
-// The Moderator owns closing the tick
-// ---------------------------------------------------------------------------
-
-test("the Moderator decides to close the tick, and says so as data", async () => {
-  const { createModeratorPersona } = await import(
-    "../../packages/runtime/src/personas/moderator/persona.js"
-  );
-  const { TickPhases } = await import(
-    "../../packages/runtime/src/personas/_shared/tick-state-machine.mts"
-  );
-  const moderator = createModeratorPersona({ clock: () => "fixed" });
-
-  const plan = moderator.advance({
-    phase: TickPhases.SUMMARIZE,
-    event: "plan_tick_close",
-    payload: {},
-    tick: 1,
-  })?.tickClose;
-
-  assert.ok(plan, "the Moderator must answer plan_tick_close");
-  assert.equal(plan.advanceTick, true);
-  assert.equal(
-    plan.recomputeAffinityField,
-    true,
-    "advancing and recomputing describe the same instant, so they travel together",
-  );
-});
-
-test("a paused Moderator closes nothing — neither the tick nor the field", async () => {
-  const { createModeratorPersona } = await import(
-    "../../packages/runtime/src/personas/moderator/persona.js"
-  );
-  const { TickPhases } = await import(
-    "../../packages/runtime/src/personas/_shared/tick-state-machine.mts"
-  );
-  const moderator = createModeratorPersona({ clock: () => "fixed" });
-  moderator.advance({ phase: TickPhases.INIT, event: "start", payload: {}, tick: 0 });
-  moderator.advance({ phase: TickPhases.OBSERVE, event: "pause", payload: {}, tick: 1 });
-
-  const plan = moderator.advance({
-    phase: TickPhases.SUMMARIZE,
-    event: "plan_tick_close",
-    payload: {},
-    tick: 1,
-  })?.tickClose;
-
-  assert.equal(plan.advanceTick, false, "a paused tick must not advance");
-  assert.equal(
-    plan.recomputeAffinityField,
-    false,
-    "and must not publish a field for a tick that never happened",
   );
 });
